Align graphical elements to center of viewport
Cleans up some of the graphical elements for ttt.js. The viewport information can be queried from main.js and is returned in a simple array object type. For better tracking of grid unit metadata, a new Box object was added to wrap both the drawing and storage of X and O values. Some modifications were made to the main Tic-tac-toe object to better handle the movement and tracking of the selector.